Rename addRing.js to mapAction
